requirements.txt
aiofiles==23.2.1
alembic==1.13.1
anyio==4.9.0
bcrypt==3.2.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
Jinja2==3.1.3
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.13
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.7
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.13.0
uvicorn==0.27.1
